Added new cluster content extraction algorithm
