+++ .gitignore
@@ -5,6 +5,9 @@
 *.tmproj
 tmtags
 
+## NETBEANS
+nbproject
+
 ## EMACS
 *~
 \#*
