+++ test/corpus/readers/tree_expected.yaml
@@ -1,0 +1,55 @@
+--- 
+:bbcnews: |-
+  A UK charity is dealing with an increasing number of young gay Muslims becoming homeless after fleeing forced marriages and so-called honour violence.
+  During a weekly drop-in group held by the Albert Kennedy Trust in London, Suni, a 20-year-old London student, helps himself to a warm mince pie and a steaming cup of coffee. 
+  In 2008, during a holiday to Pakistan to visit relatives, his parents suspected the truth about his sexuality. They believed marriage would "cure" him of what they considered to be a psychological disorder. 
+  Name 'blackened'
+  "They told me I'm going to be forced into marriage and they're looking for a girl and I'll be married in two to three months and I won't be able to come back to London," Suni said. 
+  When he refused, he was imprisoned in his family's ancestral home in a remote village of Pakistan and subjected to regular beatings and abuse as he had brought "shame" on the strict Muslim family. 
+   I think I'd be vulnerable if people knew about me - I've heard a lot of remarks in the past about people saying that gay people should die for religious reasons 
+  "I stayed there for three months and he was always beating me. He was telling me I had blackened our family name and he was saying it's a sin. I know it was just for honour." 
+  Suni managed to escape and return to the UK, penniless and homeless. 
+  Relatives and friends were reluctant to help him due to fear of violent reprisals from his family. 
+  After a night spent in a police cell, he was put in touch with the trust, which helped find him safe accommodation. 
+  'Gay demons'
+  Trust worker Annie Southerst said in the past six months there has been an increase in the number of Muslims coming to them for help. 
+  "They face threats of physical violence, actual violence and restriction of liberties," she said. 
+  "We've had people chased out of the house with knives and we have had issues around young people who had exorcisms planned to get rid of the gay demons, I suppose. 
+  "They come to us because they're homeless, or in danger of being homeless imminently. We sort out emergency accommodation for them. 
+  "But the biggest loss they face is the loss of their families. 
+  "I can't imagine what it must be like to suddenly in your late teens, early 20s suddenly not to have a family anymore." 
+  Using laws introduced by the government in November 2008, the charity has taken out four Forced Marriage Protection orders in the past few months. 
+   I'm proud to be a Muslim... and I'm proud to be gay as well - unfortunately a lot of parents don't see that 
+  The orders were introduced after ministers dropped plans to make forcing someone to marry a crime. 
+  More than 80 have been imposed so far. Breaching one is contempt of court and can carry a two-year jail term. 
+  Fazal Mahmood runs a support group for South Asian and Middle Eastern gay men, called Himat, which means strength in Urdu. 
+  "I've got about 150 people on my mail out list. 
+  "About 80%... have been coerced into marriage or been forced into marriage or are being forced into marriage," he said. 
+  Mr Mahmood says homosexuality is considered a taboo issue within close-knit Muslim communities in areas such as London, Bradford and Manchester. 
+  "I'm proud to be a Muslim, I'm proud to be South Asian, Pakistani and I'm proud to be gay as well. 
+  "Unfortunately a lot of parents don't see that. All they see is 'what is my community going to feel like when they find out my son or daughter is gay?'." 
+  In fact he advises young gay Muslims not to come out to their families. 
+  "Once you've told your family and friends about your sexuality, the next unfortunate step for your family to do is ask you to leave." 
+  Shelim, 21, lives in east London with his large Bangladeshi family, and has decided to keep his sexuality a secret. 
+  "When they do find out, they're basically going to go against it. 
+  "My relationship with them is not going to be the same, the respect they have for me is going to be different and I'm going to miss that relationship," he said. 
+  He is also worried about the repercussions within the local community if they discover he is gay. 
+  "You see people being killed for being gay and stuff. I think I'd be vulnerable if people knew about me. 
+  "I've heard a lot of remarks in the past about people saying that gay people should die for religious reasons." 
+  A special government unit tackles the issue of forced marriages. Every year it deals with around 1,600 cases of forced marriage. Three-quarters of all calls are from people of South Asian origin. 
+  Department head Olaf Henricson-Bell said gay and lesbian youngsters were particularly vulnerable. 
+  "A few weeks ago, an individual got in touch with the unit to say he'd been taken to Pakistan, forced to marry against his will, brought back to the UK then denounced by both his new wife and his family for his sexuality. 
+  "He'd been subject to physical and other abuse. When he rang us he was scared to leave the home and we had to secure police protection. 
+  "Forced marriage by its nature is an underground practice and the cases often go unreported. 
+  "The individuals involved may be reluctant to mention sexuality when they ring us or when they bring their case to the attention of the authorities," he said. 
+  The unit plans to work with the trust to produce a training programme for Lesbian, Gay, Bisexual, and Transgender organisations working with young people at risk of being forced into marriage. 
+  You can hear more at 1230 and 1800 BST on the BBC's Asian Network Reports or via the BBC iPlayer. 
+   Bookmark with:
+  * Delicious 
+  * Digg 
+  * reddit 
+  * Facebook 
+  * StumbleUpon 
+  What are these?
+   E-mail this to a friend 
+   Printable version
